+++ titiler/endpoints/factory.py
@@ -37,12 +37,7 @@ from ..dependencies import (
 )
 from ..models.mapbox import TileJSON
 from ..models.OGC import TileMatrixSetList
-from ..resources.enums import (
-    ImageType,
-    MimeTypes,
-    OptionalHeaders,
-    PixelSelectionMethod,
-)
+from ..resources.enums import ImageType, MediaType, OptionalHeader, PixelSelectionMethod
 from ..resources.responses import GeoJSONResponse, XMLResponse
 from ..templates import templates
 
@@ -109,7 +104,7 @@ class BaseTilerFactory(metaclass=abc.ABCMeta):
     gdal_config: Dict = field(default_factory=dict)
 
     # add additional headers in response
-    optional_headers: List[OptionalHeaders] = field(default_factory=list)
+    optional_headers: List[OptionalHeader] = field(default_factory=list)
 
     def __post_init__(self):
         """Post Init: register route and configure specific options."""
@@ -329,7 +324,8 @@ class TilerFactory(BaseTilerFactory):
                             **dataset_params.kwargs,
                             **kwargs,
                         )
-                        colormap = colormap or getattr(src_dst, "colormap", None)
+                        dst_colormap = getattr(src_dst, "colormap", None)
+
             timings.append(("dataread", round(t.elapsed * 1000, 2)))
 
             if not format:
@@ -346,17 +342,18 @@ class TilerFactory(BaseTilerFactory):
                 content = image.render(
                     add_mask=render_params.return_mask,
                     img_format=format.driver,
-                    colormap=colormap,
+                    colormap=colormap or dst_colormap,
                     **format.profile,
+                    **render_params.kwargs,
                 )
             timings.append(("format", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
 
-            return Response(content, media_type=format.mimetype, headers=headers)
+            return Response(content, media_type=format.mediatype, headers=headers)
 
     def tilejson(self):  # noqa: C901
         """Register /tilejson.json endpoint."""
@@ -518,9 +515,9 @@ class TilerFactory(BaseTilerFactory):
                     "tms": tms,
                     "title": "Cloud Optimized GeoTIFF",
                     "layer_name": "cogeo",
-                    "media_type": tile_format.mimetype,
+                    "media_type": tile_format.mediatype,
                 },
-                media_type=MimeTypes.xml.value,
+                media_type=MediaType.xml.value,
             )
 
     ############################################################################
@@ -557,7 +554,7 @@ class TilerFactory(BaseTilerFactory):
                         )
             timings.append(("dataread", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 response.headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
@@ -616,15 +613,16 @@ class TilerFactory(BaseTilerFactory):
                     img_format=format.driver,
                     colormap=colormap,
                     **format.profile,
+                    **render_params.kwargs,
                 )
             timings.append(("format", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
 
-            return Response(content, media_type=format.mimetype, headers=headers)
+            return Response(content, media_type=format.mediatype, headers=headers)
 
     ############################################################################
     # /crop (Optional)
@@ -683,15 +681,16 @@ class TilerFactory(BaseTilerFactory):
                     img_format=format.driver,
                     colormap=colormap,
                     **format.profile,
+                    **render_params.kwargs,
                 )
             timings.append(("format", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
 
-            return Response(content, media_type=format.mimetype, headers=headers)
+            return Response(content, media_type=format.mediatype, headers=headers)
 
 
 @dataclass
@@ -1140,18 +1139,19 @@ class MosaicTilerFactory(BaseTilerFactory):
                     img_format=format.driver,
                     colormap=colormap,
                     **format.profile,
+                    **render_params.kwargs,
                 )
             timings.append(("format", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
 
-            if OptionalHeaders.x_assets in self.optional_headers:
+            if OptionalHeader.x_assets in self.optional_headers:
                 headers["X-Assets"] = ",".join(data.assets)
 
-            return Response(content, media_type=format.mimetype, headers=headers)
+            return Response(content, media_type=format.mediatype, headers=headers)
 
     def tilejson(self):  # noqa: C901
         """Add tilejson endpoint."""
@@ -1311,9 +1311,9 @@ class MosaicTilerFactory(BaseTilerFactory):
                     "tms": tms,
                     "title": "Cloud Optimized GeoTIFF",
                     "layer_name": "cogeo",
-                    "media_type": tile_format.mimetype,
+                    "media_type": tile_format.mediatype,
                 },
-                media_type=MimeTypes.xml.value,
+                media_type=MediaType.xml.value,
             )
 
     ############################################################################
@@ -1359,7 +1359,7 @@ class MosaicTilerFactory(BaseTilerFactory):
                         )
             timings.append(("dataread", round((t.elapsed - mosaic_read) * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 response.headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )